Blackjack code simplification
Code simplification and error fixes

+++ GameTime/Models/Card.cs
@@ -2,8 +2,8 @@
 {
     class Card
     {
-        public int MaxValue { get; set; }
-        public int MinValue { get; set; }
+        public byte MaxValue { get; set; }
+        public byte MinValue { get; set; }
         public string Symbol { get; set; }
     }
 }
